+++ types/components/molecules/menu.d.ts
@@ -1,0 +1,8 @@
+// Types
+import type { Dispatch, SetStateAction } from 'react'
+
+export type MenuProps = {
+	lng: string
+	menuParent?: string
+	setMenuParent?: Dispatch<SetStateAction<string>>
+}