+++ types/components/molecules/settings.d.ts
@@ -1,0 +1,8 @@
+// Types
+import type { Dispatch, SetStateAction } from 'react'
+
+export type SettingsProps = {
+	lng: string
+	menuParent?: string
+	setMenuParent?: Dispatch<SetStateAction<string>>
+}