+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release notes
 All notable changes to this project will be documented in this file.
 
+- Created an `Appbar` component for the **mobile**
+
 ## 0.2.0
 - Added **PWA** for project
 - Created an `Appbar` component for the **desktop**